Deployment of Solutions From Unified Rmd File and of Exercise from Reduced Rmd Source

This function takes as input the unified Rmd source files containing exercise problem questions and solutions to these problems. From the unified Rmd file the solution pdf document is produced. The solution section in the unified Rmd source file must be tagged by a given html-comment. These tags are used to produce a reduced Rmd source file which does not contain the solutions to the exercise problems. The reduced Rmd file is used to produce the exercise pdf document. The unified Rmd source file contains also augmented information which is commented out. For the notebook this information will be activated. The augmented information consists of hints to the students on how to solve the exercise problems.

deploy_src_to_ex_sol(
  ps_uni_src_path,
  ps_ex_out_dir,
  ps_sol_out_dir,
  ps_nb_src_dir,
  ps_nb_out_dir,
  ps_rexpf_trg = NULL,
  pl_master_solution_tags = list(start = "master-solution-start", end =
    "master-solution-end"),
  pl_aug_info_tags = list(start = "your-solution-start", end = "your-solution-end"),
  pb_keep_src = FALSE,
  pb_debug = FALSE,
  pobj_rtt_logger = NULL
)

Arguments

ps_uni_src_path
path to unified Rmd source file

ps_ex_out_dir
output directory for exercise pdf-file

ps_sol_out_dir
output directory for solution pdf-file

ps_nb_src_dir
source directory for exercise notebook

ps_nb_out_dir
output directory for html-version of exercise notebook

ps_rexpf_trg
target directory for material related to r-exercise platform

pl_master_solution_tags
list of tags indicating section of master solution

pl_aug_info_tags
list of tags indicating section with augmented information

pb_keep_src
flag to keep sources

pb_debug
flag for producing debug output

pobj_rtt_logger
logger object

Details

Examples

if (FALSE) {
s_uni_src_path <- system.file("extdata", "asm_ex06.Rmd", package = "rteachtools")
s_ex_src_name <- tools::file_path_sans_ext(basename(s_uni_src_path))
deploy_src_to_ex_sol(ps_uni_src_path = s_uni_src_path,
                     ps_ex_out_dir = file.path("docs","ex"),
                     ps_sol_out_dir = file.path("docs","sol"),
                     ps_nb_src_dir = file.path("nb", s_ex_src_name),
                     ps_nb_out_dir = file.path("docs", "nb"))
}